functions/pyproject.toml
[tool.poetry]
name = "functions"
version = "1.0.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false
packages = [{include = "functions"}]

[tool.poetry.dependencies]
python = "^3.11"
firebase-admin = "6.5.0"
google-cloud-storage = "2.19.0"
google-cloud-firestore = "2.21.0"
google-cloud-vision = "3.10.2"
google-cloud-texttospeech = "2.29.0"
google-cloud-speech = "2.33.0"
google-cloud-aiplatform = "1.113.0"
functions-framework = "3.9.2"
Pillow = "11.3.0"
razorpay = "1.4.2"
requests = "2.32.5"
opencv-python = "4.12.0.88"
pytest = "8.4.2"
black = "23.7.0"
flake8 = "6.1.0"

[tool.poetry.dev-dependencies]
pytest-mock = "3.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
